created the data for all entities

+++ src/main/java/com/codecool/flight_api_project/airplane/Airplane.java
@@ -22,7 +22,7 @@ public class Airplane
     private Long id;
     private String manufacturer;
     private String model;
-    private int numberOfSeats;
+    private Long numberOfSeats;
     private double speed;
 
     @ManyToOne
